requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile pyproject.toml
#
bcrypt==4.0.1
    # via paramiko
boltons==23.0.0
    # via dkdeputils (pyproject.toml)
certifi==2023.5.7
    # via requests
cffi==1.15.1
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.1.0
    # via requests
cryptography==41.0.0
    # via paramiko
decorator==5.1.1
    # via fabric
fabric==3.1.0
    # via dkdeputils (pyproject.toml)
idna==3.4
    # via requests
invoke==2.1.2
    # via fabric
paramiko==3.2.0
    # via fabric
pycparser==2.21
    # via cffi
pynacl==1.5.0
    # via paramiko
pyyaml==6.0
    # via dkdeputils (pyproject.toml)
requests==2.31.0
    # via dkdeputils (pyproject.toml)
urllib3==2.0.2
    # via requests
